# src/docs/swagger.yaml
basePath: /
definitions:
  models.RegisterRequest:
    properties:
      clientSalt:
        type: string
      key:
        type: string
      username:
        type: string
    type: object
  models.UserResponse:
    properties:
      clientSalt:
        type: string
    type: object
  models.VaultResponse:
    properties:
      modifiedAt:
        type: string
    type: object
info:
  contact: {}
  description: Credentials synchronization server for KeyLox.
  license:
    name: MIT
    url: https://mit-license.org/
  title: KeyLox Server API
  version: "1.0"
paths:
  /register:
    post:
      consumes:
      - application/json
      description: Upload a vault
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Upload a vault
      tags:
      - vault
  /user/{username}:
    get:
      description: Gets a specific user
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Gets a specific user
      tags:
      - user
  /vault/{id}:
    get:
      description: Get a vault by ID
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.VaultResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get a vault by ID
      tags:
      - vaults
swagger: "2.0"